Fix getIntention() with notify=True.
Calling getIntention() with notify=True resulted in an error as the
default value of options was a tuple but notifications expect a single
string. Changing the default value of options to None and adding a check
to replace it with the appropriate default value based on the value of
notify.